fix another warning with clang

+++ include/ctfa/include/ctfa/container/set.hpp
@@ -147,7 +147,7 @@ public:
 		while (count > 0) {
 			it = first;
 			size_t step = count / 2;
-			std::advance(it, step);
+			std::advance(it, static_cast<long>(step));
 			if (*it < obj) {
 				first = ++it;
 				count -= step + 1;
